Práctica módulo 20
- Corrección de las animaciones
- Creación de materiales
- Clase base para los enemigos
