+++ TowerDefense/Assets/_Scripts/AdminToques.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67412515f25acb245adb319147ee0ed3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
