+++ TowerDefense/Assets/_Scripts/AdminUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 468b25aec92cc644d88833eca108789b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
