+++ TowerDefense/Assets/_Scripts/Enemigo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8445cf916de1a74a9132c46ddcd4fc4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
